fix(clippy): Address HIGH and MEDIUM risk warnings
**High Risk Fixes:**
- Fixed HashMap double-lookup race condition in engine.rs:274
  Changed `contains_key()` + `insert()` to atomic `entry().or_insert_with()`
- Fixed 4 needless borrows in common.rs by removing unnecessary `&` from query references

**Medium Risk Fixes (6 total):**
1. Removed unnecessary `.clone()` on Copy type FailureStrategy (stream_job_server.rs:1495)
2. Replaced manual `% 2 == 0` with idiomatic `is_multiple_of()` (processor_registry.rs:170)
3. Replaced `map_or(false, closure)` with more idiomatic `is_some_and()` (partitioner_selector.rs:279)
4. Changed length check from `len() > 0` to `!is_empty()` (partitioner_selector.rs:293)
5. Removed unnecessary let-and-return pattern (engine.rs:1231-1241)
6. Fixed missing doc comment separator (stream_job_server.rs:1606)

**Auto-fixes Applied:**
- Ran `cargo clippy --fix` for auto-fixable suggestions across lib, tests, and examples
- Auto-fixed ~130 unused imports and other mechanical issues

**Test Results:**
- All 595 unit tests pass
- All 2340+ comprehensive tests pass
- No compilation errors
- Code formatting verified

Remaining warnings are LOW RISK patterns (test utilities, dead code) that don't affect functionality.

+++ src/velostream/server/v2/sticky_partition_strategy.rs
@@ -255,7 +255,7 @@ mod tests {
         for i in 0..8 {
             let record = HashMap::new();
             let mut record_obj = StreamRecord::new(record);
-            record_obj.partition = i as i32;
+            record_obj.partition = i;
             let _partition = strategy
                 .route_record(&record_obj, &routing_context)
                 .unwrap();
